refactor(chat and smart links): move goto handler to it's own hook and add a dedicated smart link hook.
