+++ src/hooks/index.ts
@@ -23,3 +23,5 @@ export * from "./useSendChatRequest";
 export * from "./usePatchActions";
 export * from "./useGetUserSurvey";
 export * from "./useLinksFromLsp";
+export * from "./useGoToLink";
+export * from "./useSmartLinks";
